requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
web3>=6.11.0
requests>=2.31.0
python-binance>=1.0.19
ccxt>=4.0.0
solana>=0.30.0
plotly>=5.17.0
reportlab>=4.0.0
python-dotenv>=1.0.0

